requirements.txt
python-environ
#certifi==2023.7.22
#cffi==1.16.0
channels
channels-redis
daphne
Django~=4.2
djangorestframework
djangorestframework-simplejwt
Pillow
redis
#selenium==4.13.0
#sqlparse==0.4.4
typing_extensions==4.8.0
#urllib3==2.0.6
gunicorn
psycopg2-binary
boto3
django-storages
python-dotenv
locust
django-filter
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-psycopg2
django_opentracing
logstash-logger
kafka-python
prometheus_client
django_prometheus
django-silk
django-db-geventpool
celery

